+++ packages/svelte/src/lib/components/edges/BezierEdgeInternal.svelte
@@ -10,6 +10,7 @@
   export let id: $$Props['id'] = '';
   export let source: $$Props['source'] = '';
   export let target: $$Props['target'] = '';
+  export let type: $$Props['type'] = 'default';
   export let animated: $$Props['animated'] = undefined;
   export let selected: $$Props['selected'] = undefined;
   export let label: $$Props['label'] = undefined;
@@ -43,6 +44,7 @@
   id;
   source;
   target;
+  type;
   animated;
   selected;
   data;