+++ packages/svelte/src/lib/types/nodes.ts
@@ -1,5 +1,5 @@
 import type { ComponentType, SvelteComponent } from 'svelte';
-import type { NodeBase, NodeProps } from '@xyflow/system';
+import type { NodeBase, NodeProps as NodePropsBase } from '@xyflow/system';
 
 /**
  * The node data structure that gets used for the nodes prop.
@@ -7,13 +7,32 @@ import type { NodeBase, NodeProps } from '@xyflow/system';
  */
 export type Node<
   NodeData extends Record<string, unknown> = Record<string, unknown>,
-  NodeType extends string | undefined = string | undefined
+  NodeType extends string = string
 > = NodeBase<NodeData, NodeType> & {
   class?: string;
   style?: string;
 };
 
-export type NodeTypes = Record<string, ComponentType<SvelteComponent<NodeProps>>>;
+// @todo: currently generics for nodes are not really supported
+// let's fix `type: any` when we migrate to Svelte 5
+export type NodeProps<NodeType extends Node = Node> = NodePropsBase<NodeType> & {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  type: any;
+};
+
+export type NodeTypes = Record<
+  string,
+  ComponentType<
+    SvelteComponent<
+      NodeProps & {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        data: any;
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        type: any;
+      }
+    >
+  >
+>;
 
 export type DefaultNodeOptions = Partial<Omit<Node, 'id'>>;