+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @xyflow/svelte
 
+## 0.0.38
+
+## ⚠️ Breaking changes
+
+- `NodeProps` generic is a node and not only node data. `type $$Props = NodeProps<AppNode>`  
+
+## Patch changes
+
+- unify `Edge` and `Node` type handling
+- fix safari: prevent selection of viewport
+
 ## 0.0.37
 
 ## ⚠️ Breaking changes